ClockIn-Out
I have set up the 'Clock Hours' page to where it displays the name of the employee, current time, and "Clock In" button. Once the "Clock In" button is clicked an alert shows up to as if the employee would like to clock in at that exact time. Button changes to "Clock Out"